Removed redundant type conversion tests, moved some tests from test_python into test_jsproxy, test_pyproxy, and test_typeconversions. Also commented new failing tests.
